Create default parser and constructor for CommandLineParserOptions
